requirements.txt
fastapi
uvicorn
supabase
aiogram
httpx
python-multipart
python-dateutil